+++ src/Burnbytes/Core/CallbackFlags.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Burnbytes
+{
+    [Flags]
+    public enum CallbackFlags : uint
+    {
+        None = 0,
+        LastNotification = 1
+    }
+}
