+++ src/Burnbytes/Core/DriveStrings.cs
@@ -1,0 +1,9 @@
+﻿namespace Burnbytes
+{
+    public class DriveStrings
+    {
+        public string Letter { get; set; }
+
+        public string Name { get; set; }
+    }
+}
