+++ src/Burnbytes/Core/Native/SHFILEINFO.cs
@@ -1,0 +1,30 @@
+﻿using System;
+using System.Runtime.InteropServices;
+
+namespace Burnbytes
+{
+    [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Auto)]
+    internal struct SHFILEINFO
+    {
+        internal SHFILEINFO(bool _)
+        {
+            hIcon = IntPtr.Zero;
+            iIcon = 0;
+            dwAttributes = 0;
+            szDisplayName = "";
+            szTypeName = "";
+        }
+
+        internal IntPtr hIcon;
+
+        internal int iIcon;
+
+        internal uint dwAttributes;
+
+        [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 260)]
+        internal string szDisplayName;
+
+        [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 80)]
+        internal string szTypeName;
+    };
+}
