+++ src/Burnbytes/Core/PurgeProgressChangedEventArgs.cs
@@ -1,0 +1,20 @@
+﻿using System;
+
+namespace Burnbytes
+{
+    public class PurgeProgressChangedEventArgs : EventArgs
+    {
+        public PurgeProgressChangedEventArgs(long spaceFreed, long spaceToFree, CallbackFlags flags)
+        {
+            SpaceFreed = spaceFreed;
+            SpaceToFree = spaceToFree;
+            Flags = flags;
+        }
+
+        public long SpaceFreed { get; }
+
+        public long SpaceToFree { get; }
+
+        public CallbackFlags Flags { get; }
+    }
+}
