+++ src/Burnbytes/Core/ScanProgressChangedEventArgs.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace Burnbytes
+{
+    public class ScanProgressChangedEventArgs : EventArgs
+    {
+        public ScanProgressChangedEventArgs(long spaceUsed, CallbackFlags flags) => (SpaceUsed, Flags) = (spaceUsed, flags);
+
+
+        public long SpaceUsed { get; }
+
+        public CallbackFlags Flags { get; }
+    }
+}
