+++ src/Burnbytes/Forms/Scanner.cs
@@ -5,35 +5,39 @@ using System.Runtime.InteropServices;
 using System.Threading;
 using System.Windows.Forms;
 
-namespace Comet.UI
+namespace Burnbytes.Forms
 {
     public partial class Scanner : Form
     {
-        private Thread HandlerThread;
+        private readonly Thread HandlerThread;
 
         public Scanner()
         {
             InitializeComponent();
             LblIntro.Text = string.Format(LblIntro.Text, Preferences.SelectedDrive.Name);
-            HandlerThread = new Thread(new ThreadStart(() => {
+            HandlerThread = new Thread(new ThreadStart(() =>
+            {
                 Preferences.CleanupHandlers = new List<CleanupHandler>();
-                using (RegistryKey rKey = Registry.LocalMachine.OpenSubKey(Api.VolumeCacheStoreKey, false))
+                using (var rKey = Registry.LocalMachine.OpenSubKey(CleanupApi.VolumeCacheStoreKey, false))
                 {
-                    string[] subKeyNames = rKey.GetSubKeyNames();
+                    var subKeyNames = rKey.GetSubKeyNames();
+                    
                     // Adjust progress bar maximum to discovered handler count
                     Invoke((MethodInvoker)delegate
                     {
                         PrgScan.Maximum = subKeyNames.Length;
                     });
+
                     // Set up a dummy callback because COM stuff goes haywire for particular handlers if we supply none
-                    EmptyVolumeCacheCallBacks callBacks = new EmptyVolumeCacheCallBacks();
+                    var callBacks = new EmptyVolumeCacheCallBacks();
                     for (int i = 0; i < subKeyNames.Length; i++)
                     {
-                        CleanupHandler evp = Api.InitializeHandler(subKeyNames[i], Preferences.SelectedDrive.Letter);
+                        var evp = CleanupApi.InitializeHandler(subKeyNames[i], Preferences.SelectedDrive.Letter);
                         if (evp != null)
                         {
                             if (LblHandler.IsHandleCreated)
-                                Invoke((MethodInvoker)delegate {
+                                Invoke((MethodInvoker)delegate
+                                {
                                     LblHandler.Text = evp.DisplayName;
                                     PrgScan.Value = i + 1;
                                 });
@@ -53,11 +57,14 @@ namespace Comet.UI
                         }
                     }
                 }
-                Api.DeactivateHandlers(Preferences.CleanupHandlers);
+                CleanupApi.DeactivateHandlers(Preferences.CleanupHandlers);
+
                 // Sort handlers by priority, making sure they'll run in correct order
                 Preferences.CleanupHandlers.Sort((x, y) => y.Priority.CompareTo(x.Priority));
                 // A (stupid?) way to close the form once we are done cleaning
-                Invoke((MethodInvoker)delegate {
+                
+                Invoke((MethodInvoker)delegate
+                {
                     Close();
                 });
             }));
