+++ src/Burnbytes/Forms/Scanner.resx
