+++ src/Burnbytes/Program.cs
@@ -1,17 +1,15 @@
-﻿using System;
+﻿using Burnbytes.Forms;
+using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Windows.Forms;
 
-namespace Comet
+namespace Burnbytes
 {
     internal static class Program
     {
-        internal static string GetCurrentVersionTostring()
-        {
-            Version v = new Version(Application.ProductVersion);
-            return v.ToString(3);
-        }
+        internal static string GetCurrentVersionTostring() => new Version(Application.ProductVersion).ToString(3);
 
         /// <summary>
         /// The main entry point for the application.
@@ -21,10 +19,14 @@ namespace Comet
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
+
             if (Environment.OSVersion.Version.Build < 9200)
                 MessageBox.Show("You're running on a system older than Windows 8. Due to differences in Disk Cleanup architecture, only limited functionality will be available.", "Managed Disk Cleanup", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            List<DriveStrings> availableDrives = new List<DriveStrings>();
-            foreach (DriveInfo di in DriveInfo.GetDrives())
+
+            var availableDrives = new List<DriveStrings>();
+
+
+            foreach (var di in DriveInfo.GetDrives())
             {
                 if (di.DriveType == DriveType.Fixed)
                 {
@@ -37,23 +39,29 @@ namespace Comet
             }
             if (availableDrives.Count > 1)
             {
-                using (UI.DriveSelection driveSel = new UI.DriveSelection(availableDrives))
+                using (var driveSel = new DriveSelection(availableDrives))
                     Application.Run(driveSel);
+
                 if (Preferences.SelectedDrive == null)
                     return;
             }
             else
                 Preferences.SelectedDrive = availableDrives[0];
-            using (UI.Scanner scanner = new UI.Scanner())
+
+            using (var scanner = new Scanner())
                 Application.Run(scanner);
-            if (Preferences.CleanupHandlers == null || Preferences.CleanupHandlers.Count == 0)
-                return;
-            using (UI.HandlerChoice choice = new UI.HandlerChoice())
-                Application.Run(choice);
-            if (!Preferences.ProcessPurge)
-                return;
-            using (UI.Cleaner cleaner = new UI.Cleaner())
-                Application.Run(cleaner);
+
+            if (Preferences.CleanupHandlers != null && Preferences.CleanupHandlers.Any())
+            {
+                using (var choice = new HandlerChoice())
+                    Application.Run(choice);
+
+                if (!Preferences.ProcessPurge)
+                    return;
+
+                using (var cleaner = new Cleaner())
+                    Application.Run(cleaner);
+            }
         }
     }
 }
